+++ CaesarTests/TC_1_05_01.cs
@@ -13,7 +13,6 @@ namespace CaesarTests
         IWebDriver driver = new ChromeDriver();
         LoginPage loginPageInstance;
         WebDriverWait wait;
-        LocationWindow locationWindowInstance;
         CenterContainer groupLocationInstance;
         TopMenu topMenuInstance;
         MainPage mainPageInstance;
@@ -26,19 +25,18 @@ namespace CaesarTests
             driver.Manage().Window.Maximize();
             loginPageInstance = new LoginPage(driver);
             wait.Until((d) => LoginPage.IsLoginPageOpened(d));
-            loginPageInstance.LogIn("sasha", "1234");
+            loginPageInstance.LogIn("dmytro", "1234");
             wait.Until((d) => MainPage.IsMainPageOpened(d));
             mainPageInstance = new MainPage(driver);
 
         }
         [Test]
-        public void ExecuteTest_ChooseLocationChernivtsy_LocationPageOpened()
+        public void ExecuteTest_ChooseLocationSofia_UsingDoubleClick_LocationPageOpened()
         {
             topMenuInstance = mainPageInstance.MoveToTopMenu();
             Acts.Click(topMenuInstance.LocationsItem);
-            Acts.Click(mainPageInstance.ModalWindow.LocationWindow.CityChernivtsy);
-            Acts.Click(mainPageInstance.ModalWindow.LocationWindow.ConfurmButton);
-            string exeptualResultTitle = "Chernivtsy";
+            mainPageInstance.DoubleClick(mainPageInstance.ModalWindow.LocationWindow.CitySofia);
+            string exeptualResultTitle = "Sofia";
             groupLocationInstance = new CenterContainer(driver);
             Console.WriteLine(groupLocationInstance.GroupLocation.Text);
             Assert.AreEqual(exeptualResultTitle, groupLocationInstance.GroupLocation.Text);
