+++ CaesarTests/TC_1_05_05.cs
@@ -4,6 +4,7 @@ using OpenQA.Selenium;
 using OpenQA.Selenium.Chrome;
 using OpenQA.Selenium.Support.UI;
 using System;
+using OpenQA.Selenium.Interactions;
 
 namespace CaesarTests
 {
@@ -13,40 +14,40 @@ namespace CaesarTests
         IWebDriver driver = new ChromeDriver();
         LoginPage loginPageInstance;
         WebDriverWait wait;
+        LocationWindow locationPageInstance;
         CenterContainer groupLocationInstance;
         TopMenu topMenuInstance;
         MainPage mainPageInstance;
 
-        [OneTimeSetUp]
+        [SetUp]
         public void Initialize()
         {
-            wait = new WebDriverWait(driver, TimeSpan.FromSeconds(20));
+            wait = new WebDriverWait(driver, TimeSpan.FromSeconds(4));
             driver.Url = @"http://localhost:3000/logout";
             driver.Manage().Window.Maximize();
             loginPageInstance = new LoginPage(driver);
             wait.Until((d) => LoginPage.IsLoginPageOpened(d));
-            loginPageInstance.LogIn("dmytro", "1234");
+            loginPageInstance.LogIn("sasha", "1234");
             wait.Until((d) => MainPage.IsMainPageOpened(d));
             mainPageInstance = new MainPage(driver);
-
         }
         [Test]
-        public void ExecuteTest_ChooseLocationSofia_UsingDoubleClick_LocationPageOpened()
+        public void ExecuteTest_LocationPageCloseWithoutChoosing()
         {
-            topMenuInstance = mainPageInstance.MoveToTopMenu();
+            TopMenu topMenuInstance = mainPageInstance.MoveToTopMenu();
             Acts.Click(topMenuInstance.LocationsItem);
-            mainPageInstance.DoubleClick(mainPageInstance.ModalWindow.LocationWindow.CitySofia);
-            string exeptualResultTitle = "Sofia";
+            Acts.Click(mainPageInstance.ModalWindow.LocationWindow.CityLviv);
+            Acts.Click(mainPageInstance.ModalWindow.LocationWindow.CancelButton);
+            string exeptualResultTitle = "Dnipro";
             groupLocationInstance = new CenterContainer(driver);
             Console.WriteLine(groupLocationInstance.GroupLocation.Text);
             Assert.AreEqual(exeptualResultTitle, groupLocationInstance.GroupLocation.Text);
         }
 
-
         [OneTimeTearDown]
         public void CleanUp()
         {
-            driver.Quit();
+           driver.Quit();
         }
     }
 }
